Hotfix stalinks permissions roles (#1451)
* Create migrations and model

* Add permissions in module page

* Setup permission selection in roles

* Add basic crud in roles permissions

* Attach permissions in user object

* Update role id in console up to http

* Update role id in jobs up to mail files

* Update role id from models to services files

* Update role id in stored procs

* Update sidebar permissions

* Incorporate permissions in front end, add permissions per role

* Protect front end routes using permissions

* Add backend protection with error message

+++ app/Http/Controllers/ArticlesController.php
@@ -15,7 +15,7 @@ use Illuminate\Support\Facades\DB;
 use App\Models\Registration;
 use App\Models\Publisher;
 use App\Models\Price;
-
+use Illuminate\Support\Facades\Gate;
 
 
 class ArticlesController extends Controller
@@ -318,6 +318,10 @@ class ArticlesController extends Controller
     }
 
     public function store(Request $request, NotificationInterface $notification) {
+        if (Gate::denies('create-article-article')) {
+            abort(422, 'Unauthorized Access');
+        }
+
         $request->validate([
             'backlink' => 'required',
             'writer' => 'required',
@@ -343,6 +347,10 @@ class ArticlesController extends Controller
     }
 
     public function updateContent(Request $request, NotificationInterface $notification){
+        if (Gate::denies('update-article-article')) {
+            abort(422, 'Unauthorized Access');
+        }
+
         $user_id = Auth::user()->id;
         $article = Article::find($request->get('content')['id']);
         $price_id = $article->id_writer_price ?: null;
@@ -457,6 +465,10 @@ class ArticlesController extends Controller
     }
 
     public function deleteArticle(Request $request) {
+        if (Gate::denies('delete-admin-article-admin-article')) {
+            abort(422, 'Unauthorized Access');
+        }
+
         $article = Article::find($request->id);
         $article->delete();
